Updated to 1.1.0

+++ src/main/java/fr/rowlaxx/convertutils/converters/DoubleConverter.java
@@ -16,7 +16,7 @@ public class DoubleConverter extends SimpleConverter<Double> {
 	
 	@ConvertMethod
 	public double toDouble(String string) {
-		return Integer.parseInt(string);
+		return Double.parseDouble(string);
 	}
 	
 	@ConvertMethod
